Industrial-Digital-Twin-SCADA/requirements.txt
# ============================================================
#  Industrial-Digital-Twin-SCADA — Python 依赖清单
#  生成时间：2026-03-25
# ============================================================

# ── 通用依赖（两个子系统共用）──────────────────────────────────
paho-mqtt>=2.1.0          # MQTT 客户端，用于边缘网关与数据记录器的消息通信
Flask>=3.0.0              # 轻量级 Web 框架，提供 HMI 监控台后端服务

# ── 01_Virtual_PLC_Simulation（纯软仿真）──────────────────────
pymodbus>=3.11.0          # Modbus TCP 协议栈，用于虚拟 PLC 和边缘网关的 Modbus 通信

# ── 02_TIA_Portal_HIL（软硬件在环）────────────────────────────
python-snap7>=1.4.1       # Snap7 S7 通信库，用于直连西门子 S7-1200 PLC 的 DB 块读写
